requirements.txt
Flask==2.0.2
line-bot-sdk==2.3.0
json5==0.9.6
pandas==1.4.1
requests==2.27.1